+++ src/BayesFilters/src/Resampling.cpp
@@ -54,19 +54,19 @@ void Resampling::resample(const ParticleSet& cor_particles, ParticleSet& res_par
                           Ref<VectorXi> res_parents)
 {
     int num_particles = static_cast<int>(cor_particles.weight().rows());
-    VectorXf csw(num_particles);
+    VectorXd csw(num_particles);
 
     csw(0) = cor_particles.weight(0);
     for (int i = 1; i < num_particles; ++i)
         csw(i) = csw(i-1) + cor_particles.weight(i);
 
-    std::uniform_real_distribution<float> distribution_res(0.0, 1.0/num_particles);
-    float u_1 = distribution_res(generator_);
+    std::uniform_real_distribution<double> distribution_res(0.0, 1.0/num_particles);
+    double u_1 = distribution_res(generator_);
 
     int idx_csw = 0;
     for (int j = 0; j < num_particles; ++j)
     {
-        float u_j = u_1 + static_cast<float>(j)/num_particles;
+        double u_j = u_1 + static_cast<double>(j)/num_particles;
 
         while (u_j > csw(idx_csw) && idx_csw < (num_particles - 1))
             idx_csw += 1;
@@ -80,7 +80,7 @@ void Resampling::resample(const ParticleSet& cor_particles, ParticleSet& res_par
 }
 
 
-float Resampling::neff(const Ref<const VectorXf>& cor_weights)
+double Resampling::neff(const Ref<const VectorXd>& cor_weights)
 {
     return 1.0/cor_weights.array().square().sum();
 }
