+++ src/BayesFilters/src/SimulatedLinearSensor.cpp
@@ -7,8 +7,8 @@ using namespace Eigen;
 SimulatedLinearSensor::SimulatedLinearSensor
 (
     std::unique_ptr<SimulatedStateModel> simulated_state_model,
-    const float sigma_x,
-    const float sigma_y,
+    const double sigma_x,
+    const double sigma_y,
     const unsigned int seed
 ) :
     LinearModel(sigma_x, sigma_y, seed),
@@ -19,8 +19,8 @@ SimulatedLinearSensor::SimulatedLinearSensor
 SimulatedLinearSensor::SimulatedLinearSensor
 (
      std::unique_ptr<SimulatedStateModel> simulated_state_model,
-     const float sigma_x,
-     const float sigma_y
+     const double sigma_x,
+     const double sigma_y
 ) :
     LinearModel(sigma_x, sigma_y),
     simulated_state_model_(std::move(simulated_state_model))
@@ -42,9 +42,9 @@ bool SimulatedLinearSensor::freezeMeasurements()
     if (!simulated_state_model_->bufferData())
         return false;
 
-    measurement_ = H_ * any::any_cast<MatrixXf>(simulated_state_model_->getData());
+    measurement_ = H_ * any::any_cast<MatrixXd>(simulated_state_model_->getData());
 
-    MatrixXf noise;
+    MatrixXd noise;
     std::tie(std::ignore, noise) = getNoiseSample(measurement_.cols());
 
     measurement_ += noise;
